removed unused methods in AssetBundleLoadHelper, renamed it to DependencyMapLoadHelper
